streamline api routes
