+++ genzen-backend/src/routers/chat.py
@@ -1,0 +1,5 @@
+from fastapi import APIRouter
+import uuid
+from src.connections.db import get_connection
+
+router = APIRouter()
